Remove 7.6.3 and 7.8.4 from the travis build
The build badge says that it fails to build, looking at the travis output however, it only looks like it fails to build for ghc 7.6.3 and 7.8.4. By dropping them, the badge should hopefully show that it's building; not leaving the impression that this repo doesn't build.

---

Ideally we would have `[ build | ghc-X.Y.Z]` badges for each ghc, so the suported ghc versions would be more obvious.